refactor(tests): Extract three more test suites from monolithic file
Extract ContentViewModelTests, JournalScheduleTests, and NotificationDelegateTests
into separate files to improve test organization and maintainability.

Changes:
- Created ContentViewModelTests.swift (182 lines, 11 tests)
- Created JournalScheduleTests.swift (51 lines, 3 tests)
- Created NotificationDelegateTests.swift (116 lines, 4 tests)
- Moved shared test utilities to TestUtilities.swift:
  * SampleData enum
  * CatalogRepositoryMock and JournalClientMock
  * CatalogRemoteStub, FailingRemoteStub, CatalogRepositoryLoggerSpy
  * APIClientSpy
- Reduced monolithic file from 754 to 241 lines (~68% reduction)
- All 18 test suites passing

Test suites already in run-tests-individually.sh, no script updates needed.

Resolves #58
