+++ frontend/WavelengthWatch/WavelengthWatch Watch AppTests/WavelengthWatch_Watch_AppTests.swift
@@ -4,48 +4,6 @@ import Testing
 import UserNotifications
 @testable import WavelengthWatch_Watch_App
 
-private enum SampleData {
-  static let catalog: CatalogResponseModel = {
-    let medicinal = CatalogCurriculumEntryModel(id: 1, dosage: .medicinal, expression: "Commitment")
-    let toxic = CatalogCurriculumEntryModel(id: 2, dosage: .toxic, expression: "Overcommitment")
-    let strategy = CatalogStrategyModel(id: 3, strategy: "Cold Shower", color: "Blue")
-    let phase = CatalogPhaseModel(id: 1, name: "Rising", medicinal: [medicinal], toxic: [toxic], strategies: [strategy])
-
-    // Mimic production: Layers 0-10 (strategies + all spiral dynamics colors)
-    let layers = [
-      // Layer 0: Strategies
-      CatalogLayerModel(id: 0, color: "Strategies", title: "SELF-CARE", subtitle: "(Strategies)", phases: [phase]),
-      // Layers 1-10: Emotion layers (spiral dynamics)
-      CatalogLayerModel(id: 1, color: "Beige", title: "BEIGE", subtitle: "(Survival)", phases: [phase]),
-      CatalogLayerModel(id: 2, color: "Purple", title: "PURPLE", subtitle: "(Tribal)", phases: [phase]),
-      CatalogLayerModel(id: 3, color: "Red", title: "RED", subtitle: "(Power)", phases: [phase]),
-      CatalogLayerModel(id: 4, color: "Blue", title: "BLUE", subtitle: "(Order)", phases: [phase]),
-      CatalogLayerModel(id: 5, color: "Orange", title: "ORANGE", subtitle: "(Achievement)", phases: [phase]),
-      CatalogLayerModel(id: 6, color: "Green", title: "GREEN", subtitle: "(Community)", phases: [phase]),
-      CatalogLayerModel(id: 7, color: "Yellow", title: "YELLOW", subtitle: "(Integral)", phases: [phase]),
-      CatalogLayerModel(id: 8, color: "Turquoise", title: "TURQUOISE", subtitle: "(Holistic)", phases: [phase]),
-      CatalogLayerModel(id: 9, color: "Coral", title: "CORAL", subtitle: "(Transpersonal)", phases: [phase]),
-      CatalogLayerModel(id: 10, color: "Teal", title: "TEAL", subtitle: "(Unitive)", phases: [phase]),
-    ]
-
-    return CatalogResponseModel(phaseOrder: ["Rising"], layers: layers)
-  }()
-}
-
-final class CatalogRemoteStub: CatalogRemoteServicing {
-  var fetchCount = 0
-  var response: CatalogResponseModel
-
-  init(response: CatalogResponseModel) {
-    self.response = response
-  }
-
-  func fetchCatalog() async throws -> CatalogResponseModel {
-    fetchCount += 1
-    return response
-  }
-}
-
 struct CatalogRepositoryTests {
   private func makeRepository(
     remote: CatalogRemoteStub,
@@ -127,52 +85,6 @@ struct CatalogRepositoryTests {
   }
 }
 
-final class CatalogRepositoryLoggerSpy: CatalogRepositoryLogging {
-  private(set) var errors: [Error] = []
-
-  func cacheDecodingFailed(_ error: Error) {
-    errors.append(error)
-  }
-}
-
-final class FailingRemoteStub: CatalogRemoteServicing {
-  let error: Error
-
-  init(error: Error) {
-    self.error = error
-  }
-
-  func fetchCatalog() async throws -> CatalogResponseModel {
-    throw error
-  }
-}
-
-final class APIClientSpy: APIClientProtocol {
-  var lastPath: String?
-  var lastBody: Data?
-  var response = JournalResponseModel(id: 10, curriculumID: 5, secondaryCurriculumID: 7, strategyID: 9, initiatedBy: .self_initiated)
-  private let encoder: JSONEncoder
-  private let decoder: JSONDecoder
-
-  init() {
-    self.encoder = JSONEncoder()
-    encoder.dateEncodingStrategy = .iso8601
-    self.decoder = JSONDecoder()
-    decoder.dateDecodingStrategy = .iso8601
-  }
-
-  func get<T>(_ path: String) async throws -> T where T: Decodable {
-    throw NSError(domain: "unimplemented", code: 1)
-  }
-
-  func post<Response>(_ path: String, body: some Encodable) async throws -> Response where Response: Decodable {
-    lastPath = path
-    lastBody = try encoder.encode(body)
-    let data = try encoder.encode(response)
-    return try decoder.decode(Response.self, from: data)
-  }
-}
-
 struct JournalClientTests {
   @Test func encodesPayloadWithStableUserIdentifier() async throws {
     let spy = APIClientSpy()
@@ -195,388 +107,6 @@ struct JournalClientTests {
     #expect(payload.initiatedBy == .scheduled)
     let expected = Int("123456781234", radix: 16) ?? 0
     #expect(payload.userID == expected)
-  }
-}
-
-final class CatalogRepositoryMock: CatalogRepositoryProtocol {
-  var cached: CatalogResponseModel?
-  var result: Result<CatalogResponseModel, Error>
-  var loadCalls = 0
-  var lastForceRefresh: Bool?
-
-  init(cached: CatalogResponseModel? = nil, result: Result<CatalogResponseModel, Error>) {
-    self.cached = cached
-    self.result = result
-  }
-
-  func cachedCatalog() -> CatalogResponseModel? {
-    cached
-  }
-
-  func loadCatalog(forceRefresh: Bool) async throws -> CatalogResponseModel {
-    loadCalls += 1
-    lastForceRefresh = forceRefresh
-    return try result.get()
-  }
-
-  func refreshCatalog() async throws -> CatalogResponseModel {
-    try await loadCatalog(forceRefresh: true)
-  }
-}
-
-final class JournalClientMock: JournalClientProtocol {
-  struct ErrorStub: Error {}
-
-  var submissions: [(Int, Int?, Int?)] = []
-  var submittedInitiatedBy: InitiatedBy?
-  var shouldFail = false
-
-  func submit(
-    curriculumID: Int,
-    secondaryCurriculumID: Int?,
-    strategyID: Int?,
-    initiatedBy: InitiatedBy
-  ) async throws -> JournalResponseModel {
-    submissions.append((curriculumID, secondaryCurriculumID, strategyID))
-    submittedInitiatedBy = initiatedBy
-    if shouldFail {
-      throw ErrorStub()
-    }
-    return JournalResponseModel(id: 1, curriculumID: curriculumID, secondaryCurriculumID: secondaryCurriculumID, strategyID: strategyID, initiatedBy: initiatedBy)
-  }
-}
-
-struct ContentViewModelTests {
-  @Test func loadsCatalogSuccessfully() async throws {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    #expect(!viewModel.layers.isEmpty)
-    #expect(viewModel.loadErrorMessage == nil)
-    #expect(viewModel.isLoading == false)
-  }
-
-  @Test func surfacesErrorWhenLoadingFails() async {
-    enum TestError: Error { case failure }
-    let repository = CatalogRepositoryMock(result: .failure(TestError.failure))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    #expect(viewModel.layers.isEmpty)
-    #expect(viewModel.loadErrorMessage != nil)
-  }
-
-  @Test func retriesAfterFailure() async {
-    enum TestError: Error { case failure }
-    let repository = CatalogRepositoryMock(result: .failure(TestError.failure))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-    #expect(viewModel.layers.isEmpty)
-
-    repository.result = .success(SampleData.catalog)
-    await viewModel.retry()
-
-    #expect(!viewModel.layers.isEmpty)
-    #expect(viewModel.loadErrorMessage == nil)
-  }
-
-  @Test func reportsJournalOutcome() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.journal(curriculumID: 1)
-    switch viewModel.journalFeedback?.kind {
-    case .success?: #expect(true)
-    default: #expect(Bool(false))
-    }
-
-    journal.shouldFail = true
-    await viewModel.journal(curriculumID: 1)
-    switch viewModel.journalFeedback?.kind {
-    case let .failure(message)?:
-      #expect(message.contains("try again"))
-    default: #expect(Bool(false))
-    }
-  }
-
-  @Test func logsJournalEntriesWithStrategy() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.journal(curriculumID: 1, strategyID: 3)
-
-    #expect(journal.submissions.count == 1)
-    let submission = journal.submissions[0]
-    #expect(submission.0 == 1) // curriculumID
-    #expect(submission.1 == nil) // secondaryCurriculumID
-    #expect(submission.2 == 3) // strategyID
-  }
-
-  @Test func logsJournalEntriesWithSecondaryCurriculum() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.journal(curriculumID: 1, secondaryCurriculumID: 2)
-
-    #expect(journal.submissions.count == 1)
-    let submission = journal.submissions[0]
-    #expect(submission.0 == 1) // curriculumID
-    #expect(submission.1 == 2) // secondaryCurriculumID
-    #expect(submission.2 == nil) // strategyID
-  }
-
-  // MARK: - Layer Filtering Tests
-
-  @Test func filteredLayersDefaultsToAll() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    // Default filter mode should be .all
-    #expect(viewModel.layerFilterMode == .all)
-    #expect(viewModel.filteredLayers.count == viewModel.layers.count)
-  }
-
-  @Test func filteredLayersWithEmotionsOnlyExcludesLayerZero() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    // Set filter to emotions only
-    viewModel.layerFilterMode = .emotionsOnly
-
-    // Should exclude layer 0 (11 total layers, 10 emotion layers)
-    #expect(!viewModel.filteredLayers.contains { $0.id == 0 })
-    #expect(viewModel.filteredLayers.count == 10)
-    #expect(viewModel.filteredLayers.count == viewModel.layers.count - 1)
-  }
-
-  @Test func filteredLayersWithStrategiesOnlyIncludesOnlyLayerZero() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    // Set filter to strategies only
-    viewModel.layerFilterMode = .strategiesOnly
-
-    // Should include only layer 0
-    #expect(viewModel.filteredLayers.count == 1)
-    #expect(viewModel.filteredLayers.first?.id == 0)
-  }
-
-  @Test func filteredLayersReactsToModeChange() async {
-    let repository = CatalogRepositoryMock(cached: SampleData.catalog, result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    await viewModel.loadCatalog()
-
-    let originalCount = viewModel.filteredLayers.count
-
-    // Change to emotions only
-    viewModel.layerFilterMode = .emotionsOnly
-    let emotionsCount = viewModel.filteredLayers.count
-
-    // Change to strategies only
-    viewModel.layerFilterMode = .strategiesOnly
-    let strategiesCount = viewModel.filteredLayers.count
-
-    // Verify filtering works (prod has 11 layers: 0-10)
-    #expect(originalCount == 11) // All layers (0-10)
-    #expect(emotionsCount == 10) // Only emotion layers (1-10)
-    #expect(strategiesCount == 1) // Only strategy layer (0)
-    #expect(originalCount > emotionsCount)
-    #expect(emotionsCount > strategiesCount)
-  }
-
-  @Test func filteredLayersWhenLayersEmptyReturnsEmpty() {
-    let repository = CatalogRepositoryMock(result: .success(SampleData.catalog))
-    let journal = JournalClientMock()
-    let viewModel = ContentViewModel(repository: repository, journalClient: journal)
-
-    // Don't load catalog, layers should be empty
-    #expect(viewModel.layers.isEmpty)
-    #expect(viewModel.filteredLayers.isEmpty)
-
-    // Try different filter modes
-    viewModel.layerFilterMode = .emotionsOnly
-    #expect(viewModel.filteredLayers.isEmpty)
-
-    viewModel.layerFilterMode = .strategiesOnly
-    #expect(viewModel.filteredLayers.isEmpty)
-  }
-}
-
-struct JournalScheduleTests {
-  @Test func encodesAndDecodesSchedule() throws {
-    var time = DateComponents()
-    time.hour = 8
-    time.minute = 0
-
-    let schedule = JournalSchedule(
-      id: UUID(),
-      time: time,
-      enabled: true,
-      repeatDays: [1, 2, 3, 4, 5]
-    )
-
-    let encoder = JSONEncoder()
-    let data = try encoder.encode(schedule)
-
-    let decoder = JSONDecoder()
-    let decoded = try decoder.decode(JournalSchedule.self, from: data)
-
-    #expect(decoded.id == schedule.id)
-    #expect(decoded.time.hour == 8)
-    #expect(decoded.time.minute == 0)
-    #expect(decoded.enabled == true)
-    #expect(decoded.repeatDays == [1, 2, 3, 4, 5])
-  }
-
-  @Test func validatesRepeatDays() {
-    var time = DateComponents()
-    time.hour = 8
-    time.minute = 0
-
-    let validSchedule = JournalSchedule(time: time, repeatDays: [0, 6])
-    #expect(validSchedule.isValid)
-
-    let invalidSchedule = JournalSchedule(time: time, repeatDays: [-1, 7])
-    #expect(!invalidSchedule.isValid)
-  }
-
-  @Test func defaultsToAllDaysEnabled() {
-    var time = DateComponents()
-    time.hour = 8
-    time.minute = 0
-
-    let schedule = JournalSchedule(time: time)
-    #expect(schedule.enabled == true)
-    #expect(schedule.repeatDays == [0, 1, 2, 3, 4, 5, 6])
-  }
-}
-
-struct NotificationDelegateTests {
-  /// Tests the core notification handling logic by calling handleNotificationResponse with a mock response.
-  /// Note: We can't easily mock UNNotificationResponse (it's a sealed class), so we test the logic
-  /// by verifying the delegate correctly parses userInfo and updates its state.
-  /// The NotificationDelegateShim integration is verified through manual testing and production usage.
-  @Test func handlesScheduledNotificationResponse() async {
-    let delegate = await MainActor.run { NotificationDelegate() }
-
-    // Create a real notification request with userInfo
-    let content = UNMutableNotificationContent()
-    content.title = "Journal Check-In"
-    content.userInfo = [
-      "scheduleId": "test-schedule-123",
-      "initiatedBy": "scheduled",
-    ]
-
-    let request = UNNotificationRequest(
-      identifier: "test-notification",
-      content: content,
-      trigger: nil
-    )
-
-    // Test the logic by simulating what handleNotificationResponse does
-    await MainActor.run {
-      let userInfo = request.content.userInfo
-
-      if let scheduleId = userInfo["scheduleId"] as? String,
-         let initiatedByString = userInfo["initiatedBy"] as? String,
-         initiatedByString == "scheduled"
-      {
-        delegate.scheduledNotificationReceived = ScheduledNotification(
-          scheduleId: scheduleId,
-          initiatedBy: .scheduled
-        )
-      }
-    }
-
-    await MainActor.run {
-      #expect(delegate.scheduledNotificationReceived?.scheduleId == "test-schedule-123")
-      #expect(delegate.scheduledNotificationReceived?.initiatedBy == .scheduled)
-    }
-  }
-
-  @Test func ignoresNonScheduledNotifications() async {
-    let delegate = await MainActor.run { NotificationDelegate() }
-
-    let content = UNMutableNotificationContent()
-    content.userInfo = [
-      "scheduleId": "test-schedule-123",
-      "initiatedBy": "self", // Not "scheduled"
-    ]
-
-    let request = UNNotificationRequest(
-      identifier: "test-notification",
-      content: content,
-      trigger: nil
-    )
-
-    // Test the filtering logic
-    await MainActor.run {
-      let userInfo = request.content.userInfo
-
-      if let scheduleId = userInfo["scheduleId"] as? String,
-         let initiatedByString = userInfo["initiatedBy"] as? String,
-         initiatedByString == "scheduled"
-      {
-        delegate.scheduledNotificationReceived = ScheduledNotification(
-          scheduleId: scheduleId,
-          initiatedBy: .scheduled
-        )
-      }
-    }
-
-    await MainActor.run {
-      #expect(delegate.scheduledNotificationReceived == nil)
-    }
-  }
-
-  @Test func clearsNotificationState() async {
-    let delegate = await MainActor.run { NotificationDelegate() }
-
-    await MainActor.run {
-      delegate.scheduledNotificationReceived = ScheduledNotification(
-        scheduleId: "test-id",
-        initiatedBy: .scheduled
-      )
-      #expect(delegate.scheduledNotificationReceived != nil)
-    }
-
-    await delegate.clearNotificationState()
-
-    await MainActor.run {
-      #expect(delegate.scheduledNotificationReceived == nil)
-    }
-  }
-
-  /// Regression test for notification delegate race condition.
-  /// Verifies that the NotificationDelegateShim has a delegate registered
-  /// immediately after app initialization, preventing dropped notifications
-  /// that could arrive before ContentView appears.
-  @Test func delegateIsRegisteredImmediately() {
-    // The shim should have a delegate set by the app's StateObject initialization
-    // This test verifies that we don't have a race condition where notifications
-    // could arrive before the delegate is registered.
-    #expect(NotificationDelegateShim.shared.delegate != nil)
   }
 }
